+++ transformers/dart-dvhcvn/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.0.20230410
+
+* 721/NQ-UBTVQH15: Nghị quyết về việc thành lập thị xã Tịnh Biên, các phường thuộc thị xã Tịnh Biên, thị trấn Đa Phước thuộc huyện An Phú và thị trấn Hội An thuộc huyện Chợ Mới, tỉnh An Giang;
+* 722/NQ-UBTVQH15: Nghị quyết về việc thành lập thị trấn Vân Tùng thuộc huyện Ngân Sơn, tỉnh Bắc Kạn;
+* 723/NQ-UBTVQH15: Nghị quyết về việc thành lập thị xã Thuận Thành và các phường thuộc thị xã Thuận Thành, thị xã Quế Võ và các phường thuộc thị xã Quế Võ, tỉnh Bắc Ninh;
+* 724/NQ-UBTVQH15: Nghị quyết về việc thành lập thị trấn Tiên Thủy thuộc huyện Châu Thành, thị trấn Tiệm Tôm thuộc huyện Ba Tri và thị trấn Phước Mỹ Trung thuộc huyện Mỏ Cày Bắc, tỉnh Bến Tre;
+* 725/NQ-UBTVQH15: Nghị quyết về việc thành lập thành phố Tân Uyên thuộc tỉnh Bình Dương;
+* 726/NQ-UBTVQH15: Nghị quyết về việc thành lập thị trấn Powng Drang thuộc huyện Krông Búk, tỉnh Đắk Lắk;
+* 727/NQ-UBTVQH15: Nghị quyết về việc thành lập các phường thuộc thị xã Điện Bàn và thị trấn Trung Phước thuộc huyện Nông Sơn, tỉnh Quảng Nam;
+* 728/NQ-UBTVQH15: Nghị quyết về việc điều chỉnh địa giới đơn vị hành chính giữa xã Long Khánh và xã Ngũ Lạc thuộc huyện Duyên Hải, tỉnh Trà Vinh;
+* 729/NQ-UBTVQH15: Nghị quyết về việc thành lập thị trấn Hóa Thượng thuộc huyện Đồng Hỷ và nhập xã Quân Chu vào thị trấn Quân Chu thuộc huyện Đại Từ, tỉnh Thái Nguyên;
+* 730/NQ-UBTVQH15: Nghị quyết về việc thành lập thị trấn Kim Long thuộc huyện Tam Dương, thị trấn Tam Hồng thuộc huyện Yên Lạc và phường Định Trung thuộc thành phố Vĩnh Yên, tỉnh Vĩnh Phúc;
+
 ## 2.0.20221001
 
 * 510/NQ-UBTVQH15: Nghị quyết thành lập thị trấn Phương Sơn thuộc huyện Lục Nam và thị trấn Bắc Lý thuộc huyện Hiệp Hòa, tỉnh Bắc Giang;